+++ solution/backend/regcore/v3views/utils.py
@@ -1,0 +1,4 @@
+from drf_spectacular.utils import OpenApiParameter
+
+def OpenApiPathParameter(name, description, type):
+    return OpenApiParameter(name=name, description=description, required=True, type=type, location=OpenApiParameter.PATH)
